Use container for field layouts.
Fixes #25
The layout for each field is now required to be a container.
This is transparent in the syntax, the written layout being
always wrapped inside the some layout depending on the field's
attributes (`required`, `multiple`).
To make this work, two new layouts have been introduced,
`option` for optional values, and `required` for required values.
The `required` container layout is a "useless" container since
it is a simple wrapper around the value, but it is necessary
to ensure the layout of a field is always a container.

I've also updated `locspan` related dependencies versions.