Generate context relatively to other contexts. Fixes #58
This is not a perfect solution, but should work most of the time.
It is also highly inefficient because of the `json-ld`
context implementation: we have to clone the input context multiple
times, and because there is no internal sharing, a lot of data might be
duplicated. If the input context is large, this may slow down the
generator.
In the future I'll introduce internal sharing in the `json-ld` context
implementation, which should solve this issue.